add check for http error

+++ src/net/CurlWrapper.hpp
@@ -10,7 +10,7 @@ static size_t callback (void* contents, size_t size, size_t nmemb, void* userp);
 class MyCurl{
 private:
 	CURL* curl;
-	long http_code = 0;
+	int http_code = 0;
 	std::string page, buffer;
 	nlohmann::json j;
